% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{nombres}
\alias{nombres}
\alias{nombres-package}
\title{El objetivo de nombres tener datos de nombres en espaniol de paises que hablen espaniol como principal lengua.
Son bienvenidas contribuciones que vengan de fuentes oficiales de cada pais. Se entiende que no todos los paises tienen los mejores datos y se publicara lo que exista en los repositorios oficiales.
Como primer pais va a ser Argentina. En el caso de argentina los datos obtenidos son obtenidos de la url a 2019-10-18
https://datos.gob.ar/dataset/otros-nombres-personas-fisicas/archivo/otros_2.1}
\description{
El objetivo de nombres tener datos de nombres en espaniol de paises que hablen espaniol como principal lengua.
Son bienvenidas contribuciones que vengan de fuentes oficiales de cada pais. Se entiende que no todos los paises tienen los mejores datos y se publicara lo que exista en los repositorios oficiales.
Como primer pais va a ser Argentina. En el caso de argentina los datos obtenidos son obtenidos de la url a 2019-10-18
https://datos.gob.ar/dataset/otros-nombres-personas-fisicas/archivo/otros_2.1
}
