% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nombres_registrados_argentina}
\alias{nombres_registrados_argentina}
\title{Nombres registrados por anio en Argentina}
\format{Un dataframe con 3 variables: \code{nombre}, \code{cantidad}, \code{anio}}
\usage{
nombres_registrados_argentina
}
\description{
Listado historico de los nombres registrados ante el Registro Nacional de las Personas desde 1922 agrupado por nombre y anio.
Ministerio del Interior, Obras Publicas y Vivienda, Secretaria del Interior. Registro Nacional de las Personas.
Los datos estan disponibles desde 1922 hasta 2015 en la version 0 del paquete.
}
\keyword{datasets}
